Move compute_weight; del 0.05 in ks_list_from_tsv
